fix abstract super class for map
